+++ app/src/test/java/com/codesoom/assignment/auth/infra/JwtUtilTest.java
@@ -1,7 +1,7 @@
-package com.codesoom.global.utils;
+package com.codesoom.assignment.auth.infra;
 
 import com.codesoom.assignment.auth.application.InvalidTokenException;
-import com.codesoom.assignment.global.utils.JwtUtil;
+import com.codesoom.assignment.auth.infra.JwtUtil;
 import io.jsonwebtoken.Claims;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
